Integer :'

+++ src/main/java/com/example/projectapi/model/Integration.java
@@ -12,7 +12,7 @@ public class Integration {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_id", nullable = false)
@@ -25,8 +25,8 @@ public class Integration {
     @Column(columnDefinition = "jsonb")
     private JsonNode details;
 
-    public Long getId() { return id; }
-    public void setId(Long id) { this.id = id; }
+    public Integer getId() { return id; }
+    public void setId(Integer id) { this.id = id; }
 
     public User getUser() { return user; }
     public void setUser(User user) { this.user = user; }
